+++ decodering-db/src/repository.rs
@@ -57,8 +57,8 @@ impl From<SecretMappingRow> for SecretMapping {
 #[derive(sqlx::FromRow)]
 pub struct ShamirRow {
     pub id: i64,
-    pub total_shares: i16,
-    pub threshold: i16,
+    pub total_shares: i32,
+    pub threshold: i32,
     pub validation_hash: Vec<u8>,
     pub created_at: i64,
 }
